+++ src/main/java/org/ahpuh/surf/user/controller/UserController.java
@@ -3,7 +3,11 @@ package org.ahpuh.surf.user.controller;
 import lombok.RequiredArgsConstructor;
 import org.ahpuh.surf.common.s3.S3Service;
 import org.ahpuh.surf.jwt.JwtAuthentication;
-import org.ahpuh.surf.user.dto.*;
+import org.ahpuh.surf.user.dto.UserDto;
+import org.ahpuh.surf.user.dto.request.UserJoinRequestDto;
+import org.ahpuh.surf.user.dto.request.UserLoginRequestDto;
+import org.ahpuh.surf.user.dto.request.UserUpdateRequestDto;
+import org.ahpuh.surf.user.dto.response.UserLoginResponseDto;
 import org.ahpuh.surf.user.service.UserService;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
