+++ src/main/java/org/ahpuh/surf/user/dto/request/UserLoginRequestDto.java
@@ -1,4 +1,4 @@
-package org.ahpuh.surf.user.dto;
+package org.ahpuh.surf.user.dto.request;
 
 import lombok.*;
