+++ src/main/java/org/ahpuh/surf/user/entity/User.java
@@ -7,7 +7,7 @@ import org.ahpuh.surf.common.entity.BaseEntity;
 import org.ahpuh.surf.follow.entity.Follow;
 import org.ahpuh.surf.like.entity.Like;
 import org.ahpuh.surf.post.entity.Post;
-import org.ahpuh.surf.user.dto.UserUpdateRequestDto;
+import org.ahpuh.surf.user.dto.request.UserUpdateRequestDto;
 import org.hibernate.annotations.SQLDelete;
 import org.hibernate.annotations.Where;
 import org.springframework.security.crypto.password.PasswordEncoder;
