+++ src/main/java/org/ahpuh/surf/user/service/UserService.java
@@ -6,9 +6,9 @@ import org.ahpuh.surf.jwt.JwtAuthentication;
 import org.ahpuh.surf.jwt.JwtAuthenticationToken;
 import org.ahpuh.surf.user.converter.UserConverter;
 import org.ahpuh.surf.user.dto.UserDto;
-import org.ahpuh.surf.user.dto.UserJoinRequestDto;
-import org.ahpuh.surf.user.dto.UserLoginResponseDto;
-import org.ahpuh.surf.user.dto.UserUpdateRequestDto;
+import org.ahpuh.surf.user.dto.request.UserJoinRequestDto;
+import org.ahpuh.surf.user.dto.request.UserUpdateRequestDto;
+import org.ahpuh.surf.user.dto.response.UserLoginResponseDto;
 import org.ahpuh.surf.user.entity.User;
 import org.ahpuh.surf.user.repository.UserRepository;
 import org.springframework.security.authentication.AuthenticationManager;
